@@ -12,7 +12,7 @@
 		}
 	],
 	"require": {
-		"php": ">=5.3",
+		"php": ">=5.4",
 		"composer/installers": "*"
 	},
 	"support": {
